Struct apple_bundles::DirectoryBundle

pub struct DirectoryBundle { /* private fields */ }

An Apple bundle backed by a filesystem/directory.

Instances represent a type-agnostic bundle (macOS application bundle, iOS application bundle, framework bundles, etc).

Implementations

impl DirectoryBundle

pub fn new_from_path(directory: &Path) -> Result<Self>

Open an existing bundle from a filesystem path.

The specified path should be the root directory of the bundle.

This will validate that the directory is a bundle and error if not. Validation is limited to locating an Info.plist file, which is required for all bundle types.

pub fn resolve_path(&self, path: impl AsRef<Path>) -> PathBuf

Resolve the absolute path to a file in the bundle.

pub fn root_dir(&self) -> &Path

The root directory of this bundle.

pub fn name(&self) -> &str

The on-disk name of this bundle.

This is effectively the directory name of the bundle. Contains the .app, .framework, etc suffix.

pub fn shallow(&self) -> bool

Whether this is a shallow bundle.

If false, content is likely in a Contents directory.

pub fn info_plist_path(&self) -> PathBuf

Obtain the path to the Info.plist file.

pub fn info_plist(&self) -> &Dictionary

Obtain the parsed Info.plist file.

pub fn info_plist_key_string(&self, key: &str) -> Result<Option<String>>

Obtain an Info.plist key as a String.

Will return None if the specified key doesn’t exist. Errors if the key value is not a string.

pub fn package_type(&self) -> BundlePackageType

Obtain the type of bundle.

pub fn display_name(&self) -> Result<Option<String>>

Obtain the bundle display name.

This retrieves the value of CFBundleDisplayName from the Info.plist.

pub fn identifier(&self) -> Result<Option<String>>

Obtain the bundle identifier.

This retrieves CFBundleIdentifier from the Info.plist.

pub fn version(&self) -> Result<Option<String>>

Obtain the bundle version string.

This retrieves CFBundleVersion from the Info.plist.

pub fn main_executable(&self) -> Result<Option<String>>

Obtain the name of the bundle’s main executable file.

This retrieves CFBundleExecutable from the Info.plist.

pub fn icon_files(&self) -> Result<Option<Vec<String>>>

Obtain filenames of bundle icon files.

This retrieves CFBundleIconFiles from the Info.plist.

pub fn files( &self, traverse_nested: bool ) -> Result<Vec<DirectoryBundleFile<'_>>>

Obtain all files within this bundle.

The iteration order is deterministic.

traverse_nested defines whether to traverse into nested bundles.

pub fn files_manifest(&self, traverse_nested: bool) -> Result<FileManifest>

Obtain all files in this bundle as a FileManifest.

pub fn nested_bundles(&self, descend: bool) -> Result<Vec<(String, Self)>>

Obtain all nested bundles within this one.

This walks the directory tree for directories that can be parsed as bundles.

If descend is true, we will descend into nested bundles and recursively emit nested bundles. Otherwise we stop traversal once a bundle is encountered.

pub fn framework_versions(&self) -> Result<Vec<String>>

Resolve the versions present within a framework.

Does not emit versions that are symlinks.

pub fn is_framework_version(&self) -> bool

Whether this bundle is a version within a framework bundle.

This is true if we are a framework bundle under a Versions directory.