Struct apple_bundles::DirectoryBundleFile

pub struct DirectoryBundleFile<'a> { /* private fields */ }

Represents a file in a DirectoryBundle.

Implementations

impl<'a> DirectoryBundleFile<'a>

pub fn absolute_path(&self) -> &Path

Absolute path to this file.

pub fn relative_path(&self) -> &Path

Relative path within the bundle to this file.

pub fn is_info_plist(&self) -> bool

Whether this is the Info.plist file.

pub fn is_main_executable(&self) -> Result<bool>

Whether this is the main executable for the bundle.

pub fn is_code_resources_xml_plist(&self) -> bool

Whether this is the _CodeSignature/CodeResources XML plist file.

pub fn is_notarization_ticket(&self) -> bool

Whether this is the CodeResources file holding the notarization ticket.

pub fn is_in_code_signature_directory(&self) -> bool

Whether this file is in the code signature directory.

pub fn symlink_target(&self) -> Result<Option<PathBuf>>

Obtain the symlink target for this file.

If None, the file is not a symlink.

pub fn metadata(&self) -> Result<Metadata>

Obtain metadata for this file.

Does not follow symlinks.

pub fn as_file_entry(&self) -> Result<FileEntry>

Convert this instance to a FileEntry.